noremove option (#76)
* noremove option

add option to allow for prereqs not to be removed in cli prereq install process

+++ .scripts/apt_prereqs_install.sh
@@ -2,13 +2,18 @@
 set -euo pipefail
 
 apt_prereqs_install() {
+  local NOREMOVE
+  NOREMOVE=${1-:"N"}
+
   sudo add-apt-repository universe \
     > /dev/null 2>&1 \
     || err "Error adding universe repository."
 
-  sudo apt-get remove docker docker-engine docker.io \
-    > /dev/null 2>&1 \
-    || true
+  if [[ "${NOREMOVE}" = "N" ]] ; then
+    sudo apt-get remove docker docker-engine docker.io \
+      > /dev/null 2>&1 \
+      || true
+  fi
 
   run_sh "$SCRIPTDIR" "apt_install" "apt-transport-https" \
     "ca-certificates" "software-properties-common" \
