+++ .scripts/yum_prereqs_install.sh
@@ -2,9 +2,14 @@
 set -euo pipefail
 
 yum_prereqs_install() {
-  sudo yum -y remove docker docker-engine docker.io \
-    > /dev/null 2>&1 \
-    || true
+  local NOREMOVE
+  NOREMOVE=${1-:"N"}
+
+  if [[ "${NOREMOVE}" = "N" ]] ; then
+    sudo yum -y remove docker docker-engine docker.io \
+      > /dev/null 2>&1 \
+      || true
+  fi
 
   run_sh "$SCRIPTDIR" "yum_install" \
     "curl" "git" "grep" "sed" "jq" \
