+++ .scripts/zypper_prereqs_install.sh
@@ -2,9 +2,14 @@
 set -euo pipefail
 
 zypper_prereqs_install() {
-  sudo zypper -n rm docker docker-engine docker.io \
-    > /dev/null 2>&1 \
-    || true
+  local NOREMOVE
+  NOREMOVE=${1-:"N"}
+
+  if [[ "${NOREMOVE}" = "N" ]] ; then
+    sudo zypper -n rm docker docker-engine docker.io \
+      > /dev/null 2>&1 \
+      || true
+  fi
 
   run_sh "$SCRIPTDIR" "zypp_install" \
     "curl" "git-core" "grep" "sed" "jq" \
